controls fixes
