+++ packages/client/src/app/components/modals/trading/orderbook/controls/Controls.tsx
@@ -1,9 +1,11 @@
-import { Dispatch, useEffect } from 'react';
+import { Dispatch, useEffect, useMemo } from 'react';
 import styled from 'styled-components';
 
 import { Trade, TradeType } from 'app/cache/trade';
 import { Item, NullItem } from 'network/shapes';
+import { Categories, CategoryKey } from './Categories';
 import { ItemBrowser } from './ItemBrowser';
+import { CONSUMABLE_TYPES } from './constants';
 
 export const Controls = ({
   controls,
@@ -60,17 +62,56 @@ export const Controls = ({
     };
   }, []);
 
+  // categorize items by type and create lookup maps
+  const categorizedItems = useMemo(() => {
+    const byCat = new Map<CategoryKey, Item[]>();
+    byCat.set('All', []);
+    byCat.set('Consumables', []);
+    byCat.set('Materials', []);
+    byCat.set('Currencies', []);
+
+    // sort all items into categories
+    for (const item of items) {
+      const type = (item.type || '').toUpperCase();
+      if (!type) {
+        byCat.get('All')!.push(item);
+        continue;
+      }
+      let key: CategoryKey = type;
+      if (CONSUMABLE_TYPES.has(type)) key = 'Consumables';
+      else if (type === 'MATERIAL') key = 'Materials';
+      else if (type === 'ERC20') key = 'Currencies';
+
+      if (!byCat.has(key)) byCat.set(key, []);
+      byCat.get(key)!.push(item);
+      byCat.get('All')!.push(item);
+    }
+
+    // sort each category alphabetically
+    for (const key of byCat.keys()) {
+      const category = byCat.get(key)!;
+      category.sort((a, b) => a.name.localeCompare(b.name));
+    }
+
+    return byCat;
+  }, [items]);
+
   /////////////////
   // RENDER
 
   return (
     <Container>
-      <ItemBrowser
+      <Categories
         items={items}
         selected={itemFilter}
         setSelected={setItemFilter}
         category={category as any}
         onCategoryChange={setCategory as any}
+      />
+      <ItemBrowser
+        items={categorizedItems.get(category as any) ?? []}
+        selected={itemFilter}
+        setSelected={setItemFilter}
       />
     </Container>
   );
@@ -83,7 +124,7 @@ const Container = styled.div`
   min-height: 0;
 
   display: flex;
-  flex-flow: column nowrap;
+  flex-flow: row nowrap;
   justify-content: flex-start;
-  overflow-y: scroll;
+  overflow-y: auto;
 `;
