+++ packages/client/src/app/components/modals/trading/orderbook/controls/constants.ts
@@ -1,0 +1,1 @@
+export const CONSUMABLE_TYPES = new Set(['FOOD', 'REVIVE', 'CONSUMABLE', 'LOOTBOX']);
